Add files via upload
Poster I've prepared as part of the assessment for the Advanced Data Analysis course. The poster sums up my quantitative analysis of household numbers domiciled in the Soccsksargen region of the Philippines. In particular, it looks at the variables (such as sex, income household, and others) that influence the number of households.